Split out Hello and ParamHello into resource files of their own.
