+++ src/ogd/apis/HelloAPI.py
@@ -13,6 +13,8 @@ from flask_restful import Resource, Api
 # import OGD libraries
 
 # import locals
+from ogd.apis.resources.Hello import Hello
+from ogd.apis.resources.ParamHello import ParamHello
 from ogd.apis.configs.ServerConfig import ServerConfig
 from ogd.apis.models.enums.ResponseStatus import ResponseStatus
 from ogd.apis.models.enums.RESTType import RESTType
@@ -22,62 +24,12 @@ class HelloAPI:
     @staticmethod
     def register(app:Flask, server_config:ServerConfig):
         api = Api(app)
-        api.add_resource(HelloAPI.Hello, '/hello')
-        api.add_resource(HelloAPI.ParamHello, '/p_hello/<name>')
+        api.add_resource(Hello, '/hello')
+        api.add_resource(ParamHello, '/p_hello/<name>')
         api.add_resource(HelloAPI.Version, '/version')
 
         HelloAPI.server_config = server_config
 
-    class Hello(Resource):
-        def get(self):
-            ret_val = APIResponse(
-                req_type = RESTType.GET,
-                val      = None,
-                msg      = "Hello! You GETted successfully!",
-                status   = ResponseStatus.OK)
-            return ret_val.AsDict
-
-        def post(self):
-            ret_val = APIResponse(
-                req_type = RESTType.POST,
-                val      = None,
-                msg      = "Hello! You POSTed successfully!",
-                status   = ResponseStatus.OK)
-            return ret_val.AsDict
-
-        def put(self):
-            ret_val = APIResponse(
-                req_type = RESTType.PUT,
-                val      = None,
-                msg      = "Hello! You PUTted successfully!",
-                status   = ResponseStatus.OK)
-            return ret_val.AsDict
-
-    class ParamHello(Resource):
-        def get(self, name):
-            ret_val = APIResponse(
-                req_type = RESTType.GET,
-                val      = None,
-                msg      = f"Hello {name}! You GETted successfully!",
-                status   = ResponseStatus.OK)
-            return ret_val.AsDict
-
-        def post(self, name):
-            ret_val = APIResponse(
-                req_type = RESTType.POST,
-                val      = None,
-                msg      = f"Hello {name}! You POSTed successfully!",
-                status   = ResponseStatus.OK)
-            return ret_val.AsDict
-
-        def put(self, name):
-            ret_val = APIResponse(
-                req_type = RESTType.PUT,
-                val      = None,
-                msg      = f"Hello {name}! You PUTted successfully!",
-                status   = ResponseStatus.OK)
-            return ret_val.AsDict
-    
     class Version(Resource):
         def get(self):
             ret_val = APIResponse(
